improve data se/de
